fix: refactor image_processor.js
split 3 files
